refactor(deps): move byoc / pytrickle dependency outside of comfystream package (#551)
