+++ pyproject.toml
@@ -9,7 +9,6 @@ version = "0.1.7"
 license = { file = "LICENSE" }
 dependencies = [
     "asyncio",
-    "pytrickle @ git+https://github.com/livepeer/pytrickle.git@v0.1.5",
     "comfyui @ git+https://github.com/hiddenswitch/ComfyUI.git@e62df3a8811d8c652a195d4669f4fb27f6c9a9ba",
     "aiortc",
     "aiohttp",
@@ -22,6 +21,9 @@ dependencies = [
 
 [project.optional-dependencies]
 dev = ["pytest", "pytest-cov", "ruff"]
+server = [
+    "pytrickle @ git+https://github.com/livepeer/pytrickle.git@v0.1.5"
+]
 
 [project.urls]
 Repository = "https://github.com/yondonfu/comfystream"
@@ -34,9 +36,6 @@ Icon = "https://raw.githubusercontent.com/livepeer/comfystream-docs/main/logo/ic
 [tool.setuptools]
 package-dir = {"" = "src"}
 packages = {find = {where = ["src", "nodes"]}}
-
-[tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
 
 [tool.ruff]
 line-length = 100
