+++ requirements.txt
@@ -1,5 +1,4 @@
 asyncio
-pytrickle @ git+https://github.com/livepeer/pytrickle.git@v0.1.5
 comfyui @ git+https://github.com/hiddenswitch/ComfyUI.git@e62df3a8811d8c652a195d4669f4fb27f6c9a9ba
 aiortc
 aiohttp
